Slit AtomSchema from SequenceSchema
