+++ modules/core/src/main/scala/lucuma/odb/api/schema/SequenceSchema.scala
@@ -3,8 +3,7 @@
 
 package lucuma.odb.api.schema
 
-import lucuma.core.model.Atom
-import lucuma.odb.api.model.{AtomModel, DereferencedSequence, PlannedTime, SequenceModel, StepModel}
+import lucuma.odb.api.model.{DereferencedSequence, PlannedTime, SequenceModel}
 import lucuma.odb.api.repo.OdbRepo
 import cats.effect.Effect
 import sangria.schema._
@@ -12,48 +11,14 @@ import sangria.schema._
 object SequenceSchema {
 
   import syntax.`enum`._
-  import PlannedTimeSchema._
-  import StepSchema.InstrumentStepType
 
-  implicit val AtomIdType: ScalarType[Atom.Id] =
-    ObjectIdSchema.idType[Atom.Id](name = "AtomId")
+  import AtomSchema._
+  import PlannedTimeSchema._
 
   implicit val EnumTypeSequenceType: EnumType[SequenceModel.SequenceType] =
     EnumType.fromEnumerated(
       "SequenceType",
       "Type of sequence, acquisition or science"
-    )
-
-  def AtomType[F[_]: Effect, D](
-    typePrefix:  String,
-    dynamicType: OutputType[D]
-  ): ObjectType[OdbRepo[F], AtomModel[StepModel[D]]] =
-    ObjectType(
-      name        = s"${typePrefix}Atom",
-      description = s"$typePrefix atom, a collection of steps that should be executed in their entirety",
-      fieldsFn    = () => fields(
-
-        Field(
-          name        = "id",
-          fieldType   = AtomIdType,
-          description = Some("Atom id"),
-          resolve     = _.value.id
-        ),
-
-        Field(
-          name        = "steps",
-          fieldType   = ListType(InstrumentStepType[F, D](typePrefix, dynamicType)),
-          description = Some("Individual steps that comprise the atom"),
-          resolve     = _.value.steps.toList
-        ),
-
-        Field(
-          name        = "time",
-          fieldType   = CategorizedTimeType[F],
-          description = Some("Time estimate for this atom's execution, the sum of each step's time."),
-          resolve     = c => PlannedTime.estimateAtom(c.value)
-        )
-      )
     )
 
   def SequenceType[F[_]: Effect, D](
